Threshod for no erosion added to NoiseChisel
Pure erosion is going to carve off sharp and small objects completely out
of the image. Therefore to avoid missing such sharp and small objects
(which have significant pixels, but not over a large area) NoiseChisel now
has a new `--noerodequant' option. Through it, the level of significance
can be specified. All pixels with a value larger than this significance
level will not be eroded.

In order to do this cleanly, the `gal_arraymanip_uchar_replace' was added
to the arraymanip library and some corrections were made in the comments of
the mesh library.

This finishes task #14139.