Pointer increment correction in Arithmetic
In Arithmetic's `where' operator, when the first popped operand was an
array, it would increment the pointer to it only if the third operand (the
conditional) had a value of 1. The increment is now taken outside of
ternary operator.

Also in the documentation, the example for where still had the old
greater-than operator (`>'), which was later (in commit 21eef31, "Letters,
not symbols for Arithmetic's conditional operators") changed to `gt'.