NoiseChisel's checksky option now saves blank pixel value
When the `--checksky' option is given to NoiseChisel it creates a
multi-extension FITS image with the detected pixels as the first layer and
the next extensions showing the progressive states of getting to the Sky
value.

When creating the first FITS extension to show the binary (0 or 1 pixel
valued image for undetected and detected pixels respectively), it ignored
the possibility of blank pixels. In other words the value to the seventh
argument of `gal_fits_array_to_file' function was set to `0'. It is now set
to 'p->anyblank', which is `1' when there are blank pixels in the image. So
the blank pixel value is passed on to the FITS header and FITS viewers will
respect those pixels.

This was not a major problem for any crash, it just made viewing the image
very hard (blank pixels would have a value of 255, while all the rest would
have 0 or 1).