Include directories in doc/forwebpage.sh
Gnulib's `gendocs.sh' script is called by `doc/forwebpage.sh' to build the
documentation in various formats. `gendocs.sh' needs to have `texinfo.tex'
that is also bootstrapped from Gnulib. However, it would not include its
hosting directory in the Gnuastro source files
(`bootstrapped/build-aux/'). So `gendocs.sh' would complain and abort. This
directory is now passed on to `gendocs.sh' so it can successfully build the
documentation.

I guess the reason that this was not noticed before is that we probably had
`texinfo.tex' within the `doc' directory when this script was mostly
written and tested. I think this was before the time that we implemented
the bootstrapping process.