Version function for library along with a test
The new `gnuastro.c' and `gnuastro/gnuastro.h' were added to the
libraries. Currently they just contain a `gal_gnuastro_version' function.
This function can be used to get the version of Gnuastro within a C
program. The `gnuastro/gnuastro.h' header is mainly for such functions that
are to do with Gnuastro its self. A test was also written for this library
function.

This test can be used as a template to add future library tests. Since this
test doesn't have any dependencies, no dependencies were defined for it for
parallel tests. However if future library tests have dependencies, their
dependencies can be included exactly like the `.sh' dependencies.

In the process, an indentation was added to the multi-line values of some
automake variables in `lib/Makefile.am' and `tests/Makefile.am' to make
them more readable.