Started the new Library chapter, with basic introductions
One of the major steps in finally closing task #13765 is to have a
documentation. In this branch (starting with this commit), we are starting
work on the new Library chapter. So far, only a very rudimentary
introduction is written, but it will slowly grow, first thoroughly
describing how to link with the libraries, and then describing all the
functions, classified by context/header, something like GSL.