Libraries are usable in C++
As part of tasks #13765 and #13786, the library header files can now be
included in C++ programs. The guidelines suggested by GNU Libtool (in the
"Writing C header files" section of the "Library tips") were followed. A
small test program (to print the version of Gnuastro) was written in C++
and is now also compiled along with the rest of the tests.

While going through the header files some cases where the line-lengths were
too long were also corrected.