CPP include directories overwritten
As a temporary fix to bug #49007, the `CPPFLAGS' variable was set in
`tests/Makefile.am' to ignore looking into the Gnulib C libraries when
compiling tests. As discussed in that bug, I don't consider this a
permanent solution, it is just a place-holder until we find a better and
more elegant solution.