Explanation on NaN (blank) values in Arithmetic
The fact that NaN values are not equal to anything (even themselves) is
important and since Arithmetic deals with blank pixels as NaN values, it
was important to explain this in the book.

Also, an interesting outcome of this property of NaN values is given: how
you can make a fully zero-valued image based on another.