Not-equal operator in Arithmetic
The conditional operators of Arithmetic where not complete! The not-equal
operator was missing. So it has been added now.

The ordering of the operators in the book and within the
`src/arithmetic/arithmetic.c' were also chaged so the lower/greater-equal
operators are now immediately after the lower/greater-than operators and
the equal and not-equal operators are also following each other.

The explanation in Arithmetic's `args.h' also didn't have the conditional
operators, so they were added.