Operator to select blank pixels in Arithmetic
With the equal operator in Arithmetic, we can choose certain pixels in the
image except for blank pixels (NaN in float types and fixed constants for
other types). But choosing blank pixels is necessary in many situations.

The problem with blank values is that in floating points, the IEEE NaN
standard defines this value as failing any conditional operator, so NaN is
not equal to its self. The only reliable way we can check for it is the
`isnan' function. For integer types the blank value is a constant (defined
by the type in Gnuastro's `lib/gnuastro/fits.h'). But the user can't use
that value manually because like all utilities, thanks to CFITSIO,
Arithmetic will detect their blank-ness and since all internal operations
are in double precision floating point type, they will be seen as NaN
values. So even if the user reads the blank value for integer types from
the FITS header, it is useless for them in Arithmetic.

To address this problem, an `isblank' operator is now added to
Arithmetic. Using C's `isnan' function, it will return 1 if the pixel is
NaN and 0 if it isn't. Because of the internal conversion, the type of
input (or value of blank pixels) is also irrelevant within Arithmetic.

This finishes task #14146.