Removed extra printf from developing time
While developing the `where' operator in Arithmetic, I had used a
`printf("\nhere\n");' in `src/arithmetic/arithmetic.c' to make sure it
reaches this point and continue the developing. However, I had forgot to
remove it and it had been pushed to the main repo. So it is being removed
with this commit.