Arithmetic's isblank operator size corrected
The `size' variable in Arithmetic's `opisblank' function was mistakenly set
at definition time, not after popping the operands. So `size' would be
given a value of 0 and the `isblank' operator was not doing anything! This
has been corrected now.

While I was developing, size was in its correct place but I hadn't added
the comment that is above it. So after implementing `isblank' and testing
it immediately before the commit, (having forgot the comment's main
argument) I took the setting of the value of size in the definition (before
popping). This shows how important this comment is, even though it is very
repetative in all the functions! While you are thinking on other issues,
you simply forget and these comments are necessary to remind you!