Order of input and conditional in -where- changed
Until now, Arithmetic's `where' operator would consider the third popped
operand to be the condition, and the second to be the array whose values
need to be changed. The problem with that initial ordering was that due to
the nature of our current very basic (single stack with no variables)
implementation of the reverse polish notation, it was very hard to do
several checks/value-changes with one command.

With this commit, the third popped operand is the input image and the
second is the comparison image. So it is possible to read the input image
once, and apply any number of `where' operators to it in one run of
Arithmetic.