Added commented debugging flags to tmpfs-config-make
Ever since we have made the dynamic libraries, it is not possible to easily
run some tools like valgrind any more, so a mere `CFLAGS="-g -O0"' was not
enough. We also need an `LDADD="-static" option at configure time so it
will build static libraries and link them statically to the executable so
it is easier to get to the root of the problem.

So now, whenever debugging is necessary, you can just uncomment the
configure line in `tmpfs-config-make' and run it.