Removed -junk- variable in pthread_barrier_init
The `pthread_barrier_init' function is used in systems that don't have this
POSIX function (in particular Mac OS X). In our implementation of it, to
avoid an unused variable warning, a `junk' integer was defined. However,
while the type of the unused variable was `pthread_barrierattr_t', the junk
variable was `int'! So some Mac OS X compilers would get a segmentation
fault or hang when running `make check' and abort.

The junk variable is now removed and attribute is just left there in the
arguments and unused in the function. So a warning might be issued by the
compiler, but atleast it will not give a segfault during `make check'.