pthread_barrier_t re-usable after pthread_barrier_destroy
Barriers can greatly simplify the process of spinning off threads and
waiting for them to finish. They allow us to define detached threads (where
no communication is needed between the threads). However, they are a
high-level pthreads construct and are thus optional in POSIX. Therefore
they are not present in some OSs. So in Gnuastro we had an internal
implementation.

It was reported by Alan Lefor that on one of his Mac OS X systems, `make
check' hangs and doens't finish. Since it could only be reproduced on his
system, it took us several days of successive tests to narrow down the
problem and find the cause.

In our implementation of pthread_barrier, (which was inspired from another
webpage), there was one hidden assumption: we won't be using the barrier
any more. So the `pthread_barrier_destroy' function would just destroy the
mutex and condition variable when the `pthread_barrier_wait' of the main
thread-spinner function was freed. However, all the separate thread
functions are also waiting behind the `pthread_barrier_wait' functions and
there is no particular order to distinguish the thread-spinner's wait
function from the separate threads. So it was possible that the
thread-spinner's wait would be freed before the threads. Therefore, those
threads that would get freed after the thread-spinner would hit an
un-initialized mutex and condition variable.

To fix this problem, a new element was added to the pthread_barrier
function to count how many condition variables have been freed (are no
longer needed). The `pthread_barrier_destroy' function (which is only
called once by the thread-spinner) now waits (using the `sleep' function)
until no more threads need the condition variables and only after they are
all done, will it destroy the condition variable and mutex. The wait is
1000 nano-seconds or 1 micro-second, and is only necessary when the
thread-spinner gets to destroy, so its effect on performance can be
negligible.

This fixes bug #49049.