Logical operators in Arithmetic
The three logical operators `and', `or' and `not' where added to
Arithmetic, the book was also updated and a test was written for the `or'
operator which is now checked with `make check'.

The explanation for the `where' operator was also made more clear.

The other operator lines in the `reversepolish' function were just adjusted
so all their function names are on the same column to help in reading.

This finishes task #14153.