Gnulib's new getprogname module adopted
Gnulib has recently (dated 2016-09-05 on its `NEWS' file) depreciated the
`progname' module, with a new `getprogname' module. This change has now
been implemented in Gnuastro: there is no more need to include `progname.h'
in a program's `main.c' and call the `set_program_name' function within the
`main' function. The program name is now found automatically by Gnulib's
'error' module.

Since Gnulib's `getprogname' module is a dependency of other Gnulib modules
that we include (`error' and `gendocs'), we don't need to explicity ask for
it in `bootstrap.conf' any more.  So in general this Gnulib update has
helped in simplifying Gnuastro and is very welcome.