bootstrapped/lib only included in libraries and utilities
Gnulib's header files in `bootstrapped/lib' would cause problems when
testing C++ programs since they were generally included in all compilation
commands from `configure.ac'! With this commit, these headers are only
included in the Makefile of utilities and libraries, not when making tests.

In general, since the bootstrapped headers and compiled library are not
installed, we should not be looking into them for tests (which the user
will be expecting to work similarly from anywhere on their system). With
this commit, the old solution to this problem in commit 2d15a91 (CPP
include directories overwritten) is now removed. This is a much more
fundamental and elegant solution.

This fixes bug #49007.