getprogname used in TEMPLATE utility
In commit 118de44 (Gnulib's new getprogname module adopted), the old Gnulib
module `progname' was removed. However, the TEMPLATE utility was
missed. The reason was that I just relied on the compilation errors to tell
me which file to fix. To make general changes like this to all utilities,
this is thus not a good way.

For commit 25d2ec5 (bootstrapped/lib only included in libraries and
utilities) I used this way: I just opened all the utility `Makefile.am's in
Emacs with one command: first CTRL-F (to prompt for file names), then
`./src/*/Makefile.am'. This opened all the files and I could switch between
them with CTRL-x (forward-arrow) and CTRL-x (back-arrow).