Better debugging in tmpfs-config-make
Until now, for debugging, `tmpfs-config-make' would set `LDFLAGS'. However,
GNU Libtool has better and higher-level enable/disable configure options
for this job: `--disable-shared' (for this case). This way, in case the
user needs to put some other values in `LDFLAGS', they can do it freely and
this option is directly given to Libtool's configuration options. The
commented explanations were also updated, inspired for the example in
Libtool's manual (in the `Static-only libraries').