Building and debugging, and Test scripts sections updated
These two sections had gone unchanged for a very long time. In particular
there was no reference to the recently added `tmpfs-config-make' and
`tests/during-dev.sh' scripts. So they were updated/edited. Most
significantly, the `Building and debugging' section was previously just
called `Building', but it was important to let people know that it also
discusses debugging (from the table of contents).