Static attribute for non-exported library functions
The libraries contain some functions that are only internal to the
library. They are not intended for outside usage, only to help in
modularity of the library code. So a `static' attribute was added to them
in the `.c' source files.

Therefore, "1. No one will be able to use unintentionally a function from
other module 2. In case of building a library the function will not be
exposed as exported function, so it won't be visible to end-user and won't
clutter up the list of exported functions." (quote from the bug report by
Vladimir Markelov).

This fixes bug #48610.