A release checklist added to facilitate releases
Since there are many things to do before a release, the
`doc/release-checklist.txt' checklist has been created. It includes all the
operations that should take place prior to making a release. Having this
file will facilitate more frequent releases.

At this stage, the file is just a place-holder, with a crude list of items
that came to my mind right now. It will be complete and tested with the 0.2
release and the beta-release a few days before that. It doesn't need to go
into the distribution, it should just be version controlled.