Libtool version defined for librarires
The GNU Libtool version is now defined for the libraries in
`lib/Makefile.am'. Note that the versioning of the libraries (which is used
by the shared library linker) is different from the main Gnuastro version:
it follows different rules.

Ultimately each release will have a general Gnuastro version that has a
one-to-one correspondance with a specific library version. The library
version of each release is also included in the NEWS file.