Gnuastro's version is a macro in the library
Until now, Gnuastro's version was reported by a function declared in
`gnuastro.h' and defined in `gnuastro.c'. The user would have to define a
variable, the function would allocate space and pass the pointer to the
user's variable, and the user would have to free the allocated space: all
very complicated!

The problem was that in Gnuastro, the version is recorded in
`config.h'. This file is internal to Gnuastro and can't be installed, so
the version couldn't directly be used from it. To fix the problem, the
version is now written in `gnuastro.h' by `lib/Makefile' from
`lib/gnuastro.h.in'. So any change that is to be added in the final
`gnuastro.h' has to be implemented in `gnuastro.h.in'. But the great thing
is that the version is now a macro and can be very easily used.

With the change of this function, the tests (which currently only report
the version) also had to be re-written.