Automatic linking script section added, array test
The introductory text for libraries was further edited and a new "Automatic
linking script" was added. This is a very convenient shell function that
can be used for compiling and linking any program using Gnuastro
functions. Maybe in the future, we can make it more general and install
it. It will make it really easy to write C programs using Gnuastro and not
having to worry about linking.

As part of the explanations, the old `versionc.c' test was replaced with
another one to be alittle more complicated: actually use some of the
library functions. Since the version reporting is done with a macro, the
`versionc' test wouldn't need any linking, which made it hard to use as a
reference for explaining in the text.