New function index, updated hosting institutions
The functions index is now separated from the other indexes so users can
easily get to the function they want. The output type of the function is
now also writen on a separate line before the function to be similar to the
actual code. The functions in `box.h' have been documented and some other
index entries were also corrected. The hosting institutions were also
updated in `THANKS' and also added in the acknowledgements section of the
book.