fits.h functions included in the book
The functions and necessary concepts for programming with `fits.h' have
been fully explained in the book. FITS is the main data storage format in
astronomy, so it was necessary to elaborate on how to work with
them. Ofcourse, these functions are just an initial release and they will
greatly evolve. During the process some of the functions and arguments were
re-named (and corrected throughout Gnuastro) so the names can be more
descriptive of what the function does. "The TEMPLATE utility" section was
also moved to the "Developing" chapter.