Arithmetic reports each new input FITS image
In large operations (where a large number of input images are needed), it
will be helpful if Arithmetic reports each image that it reads in. So by
default it does this now. This behavior can be suppressed with the
`--quiet' (`-q') option.

In the Arithmetic documentation, a blank lines was put before and after
theq first example for the `where' operator to make it more readable in
info format (like the second example).