threads.h documented, pthread_barrier usable outside
The inclusion of the `pthread_barrier' definitions and declarations in
`lib/gnuastro/threads.h' and `lib/threads.c' was not suitable for
libraries. Because the `HAVE_PTHREAD_BARRIER' macro that was used to define
them was defined in `config.h' which is not installed. So if someone's
operating system doesn't have `pthread_barrier' constructs, none of
Gnuastro's threading will work for them. So several new things were added
with this commit:

 - A new `IN_GNUASTRO_BUILD' macro is now defined in `config.h' which will
   let the headers know that they are in the pre-installation environment,
   not in the post-installation.

 - The `GAL_GNUASTRO_PTHREAD_BARRIER' was added to `lib/gnuastro.h.in' as
   another one of the installation parameters that installed headers should
   know about.

 - The `pthread_barrier' functions and structures are now only defined (in
   the `.c' and `.h' files) when `GAL_GNUASTRO_PTHREAD_BARRIER == 0'.

 - A test was added to `make check' to test the compiling and building of a
   multi-threaded program on the user's system. This can be useful as a
   tutorial for users to easily build multi-threaded programs and also
   allow users to widely test this feature easily as part of the checking
   package.

The declarations and definitions of `threads.h' are also now fully
documented within the book. In the process, some variables in the source
were also renamed to be more descriptive of their nature.