Replaced all occurances of utility to program in book
With the addition of installed libraries, the term `utility' (which is in
the full name of Gnuastro) warants a more general defintion of programs AND
libraries. However in the book, `utility' and `program' were used
interchangeably, which could cause confusion. Therefore with this commit,
all occurances of `utility' in the manual that should have been `program'
are now `program'.

This finishes task #13765.