Top level src/ directory renamed to bin/
From this commit, the top level `src/' directory is called `bin/' and all
occurances of `src/' within the Gnuastro files were also corrected. This
change was done since `src' is too generic and could cause confusion: is it
the source of the programs and utilities? How does it relate to `lib/'?

This new name was chosen based on the final installation directory of its
(main) contents: the executables of the programs within this directory are
installed in `prefix/bin/', similar to how the libraries (in the top level
`lib/' directory) are installed in `prefix/lib/'.