lib/gnuastro.h.in under version control
Through `.gitignore', we ask Git to ignore files ending in `.in'. However,
`lib/gnuastro.h.in' must not be ignored and is considered part of the
source: we use it go generate the installed `gnuastro.h' header. This had
not been recognized before since no `git clean -fxd' was done since we
created this file. So with this commit, this file is now also version
controlled.