Some updates in the book and README
The book and README file were reviewed and some changes were made to
accommodate the newly added libraries. Most notably some of the sections in
the developing chapter were moved to be more expressive for both the
programs and libraries. The two old "Internal libraries" and "headers"
sections were also removed. Until now the developing chapter was only
focused on programs.