Removed authors by directory list from AUTHORS
The `genauthors' script is in charge of building `AUTHORS'. Until now, it
would make one aggregate list of authors for all of Gnuastro and a list for
each major directory (programs, libraries, documentation). After renaming
the directory keeping the programs to `bin', Git would no longer recognize
the correct history easily. Besides that, the detailed list was getting too
long and complicated, so with this commit, only the aggregate list of
authors is stored in `AUTHORS'.