Links and CSS updates on book HTML
The `<meta name="viewport" ...>' tag was no included in the HTMLs so it
would not display nicely on smaller screens (smartphones). `doc/forwebpage'
now adds this line to each HTML.

Also the links to the bottom of the page are not relative any more, they
point to the main gnu.org server's version. The reason is this: With
relative addresses, the browser would keep the slashes in the address, so
we would end up with many slashes in the HTML names of pages afterwards and
MathJax wouldn't work, since we used relative addresses for it.

In the CSS, the page's margin was set to zero and the font size of the
menues were made larger.