A limit on book's webpage width in wide screens
On a wide screen the lines will become too large, so using CSS's `@media'
option, a limit was set on the maximum possible width. Also the 110% font
size was a little too large, after some tests, the 100% size appears to be
eaiser to read on wide monitors and narrow smartphones at the same time.