CSS for blockquotes, widescreen fonts slighly larger
The author of a quote was spaced too far away from the quote in the
original CSS and closer to the paragraph lines. With this commit, we
defined some CSS styles for the block quotes to the authors can be closer
to quotes rather than the next paragraph. However, this would decrease the
space in blockquote tags that don't have an author. So for any such
element, we also added a top margin.

Also, after a little more vewing on wide monitor screens and narrow
smartphone screens, I feel that the widescreen fonts needed to be slightly
increased for easier readability (atleast to my eyes!).