Upper limit magnitude calculation in MakeCatalog
The upper limit magnitude is a very important factor when doing photometry
on the fainter objects in an image (as explained more fully in the
book). With this commit, MakeCatalog can now produce the upper limit
magnitude for each object.

In MakeCatalog, the `upperlimit.c' and `upperlimit.h' source files were
added that are fully devoted to measuing the upper limit magnitude for each
object. The external function in `upperlimit.c' was written to not depend
on the MakeCatalog parameters structure. In this way, later it can easily
be moved with the libraries if necessary. Seven new parameters were added
to MakeCatalog for the job: `envseed', `upmask', `upmaskhdu', `upnum',
`upsclipmultip', `upsclipaccu', and `upnsigma'. In the `--help' output,
these are all classified in their own category to be more clear. Since it
involves random numbers, all the input parameters are also writen to the
comments of the output catalog for reproducibility.

In the book, the old "Depth and limiting magnitude" section was renamed to
"Quantifying data limits" and the explanations were updated to be more
consistent.

This finishes task #14167.