Non-installed headers moved to lib/
Until now, the non-installed header files were mixed with the installed
ones in `lib/gnuastro/' this made them hard to manage and would also make
the code harder to read for a user: which headers can a user also use in
their own programs and which are internal to Gnuastro?

So the internal headers are now mixed with the library `.c' source files in
`lib/' and `lib/gnuastro/' is only used for those headers that will be
installed.