Correctly report conf files checked
When `--onlydirconf' is given, the user and system configuration files are
not used. However, in the error message that is reported (even with this
option), we tell the users that we have checked the user and system
configuration files. This can cause confusion. So a check was added so when
this option is given, the user and system configuration files are not
printed.