Removed references to .xz in the book
After a good discussion on the gnu-prog-discuss mailing list, it was
decided to release Gnuastro in `.tar.gz' and `.tar.lz' formats. The first
is for backwords compatibility and stability and the second is for a robust
and accurate archival format as well as much better compression. So all
references to the `.tar.xz' format have been removed from the book to
encourage the much more robust and better `.tar.lz' format.

As suggested by Antonio Diaz Diaz (author of Lzip), the term "compression
rate" was changed to "compression ratio". The latter is the proper
technical term for what was intended.