Dirty fix to CSS first paragraph indentation
The indentation problem in the book's webpage can make it slightly hard to
read the text: we are accustomed to seeing the first paragraph of a
distinct part of text (for example under a title or within a quotation) not
having any indentation. So when this indentation is not corrected in such
cases, the text becomes unprofessional and strange to the eye: why should
the text under a title be indented? Is the title irrelevant to the text
(like an image)?

When the `<p>' tags have a parent like `.header', `dd', `li', `blockquote'
and the rest, it is very easy to write a CSS to apply the
indentation. However, unfortunately the regions between titles and the end
of the page in Texinfo's output is not within an environment. Also, between
the title and first paragraph there might be images, or index links.

A dirty solution to this problem was to manually select all paragraphs with
non-paragraph tags between them and a header. Since we can only work with
known numbers of tags in CSS, we then had to explicitly add a fixed number
of non-paragraphs between the headers and paragraphs. This is dirty because
it can be buggy and also if there is more than 10 tags between the title
and first paragraph, the first paragraph will be indented. So it is not the
best solution to the problem.

However, this dirty fix does make most of the text more readable, so we are
implementing it with this commit until a better and more general solution
is found. Maybe we can get in touch with the Texinfo developers and ask
them to put the text within one header inside of a parent tag. I will try
to write up a list of requests (as we do in the `doc/forwebpage' script)
and submit to them.