2 by 2 input matrix accounts for FITS positions
The FITS standard defines the center of the first pixel in the image to
have the position of 1.0000 and all FITS positions are defined based on
that. However, to do warping properly, we need to consider the full pixel
area, so it is necessary to shift the center first by 0.5, apply the
transformation, and then shift it by -0.5.

When the user gives a 3 by 3 matrix, it is their own responsability to
account for this shift. However, when they give a 2 by 2 matrix, we can do
the correction internally. This commit adds this capability: MakeCatalog
now does this correction internally when given a 2 by 2 matrix.

This process was actually done while I was working on task #14170 ("Align
option in image Warp"), so I have also partly included the `--align' option
with this commit, but if a user asks for it, MakeProfiles will complain
that it is not ready yet.