Pixel scale function added, angular distance renamed
`gal_wcs_pixel_scale_deg' is a new function added to Gnuastro's WCS
library. As the name suggests, it will give the pixel scale of the input
WCS structure in degrees. It has also been documented in the book.

`gal_wcs_angular_distance_deg' is the new name of the old function called
`gal_wcs_angular_distance'. That function would take the angles in radians
and also return its value in radians. Radians are very inconvenient and
uncommon in celestial coordinates, so its input and output was modified to
accept angles in degrees and also return the angular distance in
degrees. To be more explicit about the units, the `_deg' suffix was added
to the name.