ImageWarp can rotate and scale without matrix
Writing the matrix manually on the command-line can be both annoying and
buggy, even a 2 by 2 matrix. For scaling you have to check if the values
are placed properly and the two values on the diagonal are equal. For
rotation you have to calculate the sine and cosine and repeat them which is
very annoying and horribly buggy.

So with this commit two new options are added to ImageWarp: `--roate' and
`--scale'. They only take one value (the angle in degrees and the scaling
factor respectively) and create the matrix internally.

Currently only one such operations can be defined in each run of
ImageWarp. So although this commit was done in the spirit of task #13562
("Separate warp options"), it does not complete it. In that task, we
planned to allow the user to sequentially give any number of scaling,
rotation, translation and etc operations to be done on the image. But that
job is a little too complicated for now (I don't have time).

However, the infra-structure is now in place to do the work easily: the
separate matrices are read in `ui.c'. To make task #13562 a reaily, we just
have to build 3 by 3 matices for each step and apply the FITS position
correction before and after all subsequent options. A linked list can also
keep the separate values.