Minor updates in macros and configure
Internally, Autoconf uses a `HAVE_' prefix for macros that keep the results
of checks, however, we were using `HAS_' and `HAVE_' for this job. To make
things homogeneous and consistent, they were all changed to the Autoconf
style (starting with `HAVE_').

In the process, the internal `HAVE_WCSLIBVERSION' has also been changed to
`HAVE_WCSLIB_VERSION'. Also some further suggestions in the messages at the
end of steps was added.

User visible library changes:

 - `GAL_GNUASTRO_HAVE_PTHREAD_BARRIER' has replaced
   `GAL_GNUASTRO_PTHREAD_BARRIER'.