Store Git's describe in outputs using libgit2
One of the main reasons to build Gnuastro was to help/encourage
reproducibility of scientific results. When the scientific research is
defined in a version controlled system, keeping the Git commit information
that produced the data-products would be very helpful. For one example of a
pipeline for a research/paper see NoiseChisel paper's reproduction pipeline
at https://gitlab.com/makhlaghi/NoiseChisel-paper . Such systems can be
implemented in larger and smaller scales and can be very helpful/necessary
for "scientific" (reproducible) results.

Since we want to encourage such reproducible research results, this job is
automatically done within Gnuastro now. When Gnuastro's programs, or the
`gal_fits_write_keys_version' library function, are run within a Git
version controlled directory (and libgit2 is installed), they store the
commit's description in any Gnuastro output product (FITS file or ASCII
table). See the manual for more details on the format of the returned
value.

Fortunately Git comes with the libgit2 library. So we don't need to make
system calls, this library will get the commit information if Gnuastro's
tools are run within a Git version controlled directory. Libgit2 is an
optional dependency and its presence is checked at configure time. So there
is no problem if its not available.

The common FITS keywords were also organized to separte the program
specific keywords from the versioning keywords. Also the "GNU Astronomy
Utilities ..." FITS `COMMENT' was removed and the Gnuastro webpage was also
removed from the FITS output to emphasize more on the more important
keywords).

User visible library changes:

 - `gal_fits_write_keys_version' will write the commit description when
   there is a Git repository and libgit2 is installed.

 - New `gnuastro/git.h' header with `gal_git_describe' function.

 - New `GAL_GNUASTRO_HAVE_LIBGIT2' macro in `gnuastro/gnuastro.h'.

This finishes task #14172.