%zu instead of %lu for printing size_t
GNU Guix recently updated their Gnuastro build to 0.2. The build for
`i686-linux' on their system had failed because of using `%lu' for printing
`size_t' type variables. The problem is that on 32bit systems, `size_t' is
32bit (so `%u' should be used), but on 64-bit systems we should be using
`%lu'. Fortunately in C99, `printf' has the `z' modifier precisely for this
problem with `size_t'. So all occurances of printing with `lu' have been
changed to `zu'.