Corrected links on webpage
After the change of the name of the Gnuastro program list, its HTML
filename name had also changed, but this was not implemented in the top
webpage. This is now corrected. I also changed the link to my old page in
Tohoku University to my new page in Lyon Observatory.