Modular warping options in ImageWarp
ImageWarp now has options for various types of warpings: `--align',
`--sheer', `--flip', `--project', `--rotate', `--scale',
`--translate'. Each one of these options can be called on the command-line
any number of times or in the configuration files. Previously the
`--align', `--rotate' and `--scale' options could only be called once.

By default the FITS coordinates (where the center of the pixel has an
integer coordinate) are corrected with two translations at the start and
end. However, the `--nofitscorrect' is now available to stop this
correction if the user doesn't want it.

This finishes task #13562.