Correct size for output columns in Table
The `data' element of Table's `outcolumn' structure (in `bin/table/main.h')
has a `void *' type to allow various types. However, to allocate space for
the output column we were using the `sizeof' operator on it! This would
give a size of 1 and would cause a crash some systems.

In the previous commit (5c41564: New function to return the size of FITS
datatype), a new function was added to the Gnuastro FITS functions to
return the number of bytes in each datatype. With that function, the
datatype's size is now also stored in the `outcolumn' structure and the
allocation is done based on that.

This fixes bug #49347.