Fixes in CosmicCalculator's documentation
In the book, the examples at the start of the "Invoking CosmicCalculator"
section have an equal sign after the short options. This was an error and
is corrected with this commit.

Francesco Montanari kindly reported this bug, his name was thus also added
to the `THANKS' file.

In this commit, the precision of printing the values (with the
`--printparams', or `-P' option) of the densities in CosmicCalculator are
increased to 5 decimals since 3 was not enough for modern measured values.

This fixes bug #49419.