Name string for man pages in all programs
A name string was added to the man page of all the utilities. This was done
to make the man page more informative and also as part of some work to
package Gnuastro for DebianAstro. This was kindly suggested by Ole
Streicher.

Help2man's `--libtool' option was also added to avoid the `lt-' prefix in
the manpage.