Fixed segfault in breaking filename to FITS keywords
The job of `gal_fits_file_name_in_keywords' is to break the input filename
into multiple FITS keywords. It doesn't just break up the filename based on
the available space, it breaks it up based on the directories so it can be
more safely used and read. Since the last character of a string has to be
`/0', the largest position where a `/' could be placed is the second last
byte. However, we were assuming that this is the last byte, so when the `/'
fell on the last usable byte, we would have a segmentation fault. This is
fixed now.